created new controller for publishing messages
