+++ src/main/java/uk/gov/hmcts/reform/orgrolemapping/controller/WelcomeController.java
@@ -15,12 +15,8 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.hmcts.reform.orgrolemapping.domain.model.UserRequest;
 import uk.gov.hmcts.reform.orgrolemapping.domain.service.BulkAssignmentOrchestrator;
-import uk.gov.hmcts.reform.orgrolemapping.servicebus.CRDTopicPublisher;
-import uk.gov.hmcts.reform.orgrolemapping.servicebus.PublishCaseWorkerData;
-import uk.gov.hmcts.reform.orgrolemapping.servicebus.JRDTopicPublisher;
 import uk.gov.hmcts.reform.orgrolemapping.v1.V1;
 
-import java.util.Arrays;
 
 @RestController
 @Slf4j
@@ -29,16 +25,9 @@ public class WelcomeController {
 
     private BulkAssignmentOrchestrator bulkAssignmentOrchestrator;
 
-    JRDTopicPublisher jrdTopicPublisher;
-    CRDTopicPublisher crdTopicPublisher;
-
     @Autowired
-    public WelcomeController(final JRDTopicPublisher jrdTopicPublisher,
-                             final CRDTopicPublisher crdTopicPublisher,
-                             BulkAssignmentOrchestrator bulkAssignmentOrchestrator) {
+    public WelcomeController(BulkAssignmentOrchestrator bulkAssignmentOrchestrator) {
 
-        this.jrdTopicPublisher = jrdTopicPublisher;
-        this.crdTopicPublisher = crdTopicPublisher;
         this.bulkAssignmentOrchestrator = bulkAssignmentOrchestrator;
 
     }
@@ -82,28 +71,6 @@ public class WelcomeController {
                 (Math.subtractExact(System.currentTimeMillis(),startTime))
         );
         return ResponseEntity.status(response.getStatusCode()).body(response.getBody());
-    }
-
-    //This method is reserved for ASB topic testing. Need to be removed later.
-    @PostMapping(value = "/send2CRD")
-    public ResponseEntity<String> send2CRD(@RequestBody String body) {
-        log.info("Sending message 2 CRD topic");
-        PublishCaseWorkerData publishCaseWorkerData = new PublishCaseWorkerData();
-        publishCaseWorkerData.setUserIds(Arrays.asList("4b141f3c-9d8b-4eb2-932e-23fa49336CRD"));
-        crdTopicPublisher.sendMessage(publishCaseWorkerData);
-        //topicPublisher.sendMessage(body);
-        return new ResponseEntity<>("{}", HttpStatus.OK);
-    }
-
-    //This method is reserved for ASB topic testing. Need to be removed later.
-    @PostMapping(value = "/send2JRD")
-    public ResponseEntity<String> send2JRD(@RequestBody String body) {
-        log.info("Sending message 2 JRD topic");
-        PublishCaseWorkerData publishCaseWorkerData = new PublishCaseWorkerData();
-        publishCaseWorkerData.setUserIds(Arrays.asList("4b141f3c-9d8b-4eb2-932e-23fa49336JRD"));
-        jrdTopicPublisher.sendMessage(publishCaseWorkerData);
-        //topicPublisher.sendMessage(body);
-        return new ResponseEntity<>("{}", HttpStatus.OK);
     }
 
     //This method needed for the functional tests, so that RAS gets enough time to create records.
